fix oversight

+++ SketchBot/Database/ServerSettingsDB.cs
@@ -178,7 +178,7 @@ namespace Sketch_Bot
                 var welcomechannel = table["welcomechannel"] == DBNull.Value ? 0 : (ulong)table["welcomechannel"];
                 var modlogchannel = table["modlogchannel"] == DBNull.Value ? 0 : (ulong)table["modlogchannel"];
                 var xprate = table["xpmultiplier"] == DBNull.Value ? 1 : (int)table["xpmultiplier"];
-                var levelupmessages = table["LevelupMessages"] == DBNull.Value ? 1 : (ulong)table["LevelupMessages"];
+                var levelupmessages = table["LevelupMessages"] == DBNull.Value ? 1 : (int)table["LevelupMessages"];
 
                 result.Add(new Serversettings
                 {
